+++ src/decorators/InterfaceType.ts
@@ -2,14 +2,14 @@ import { getMetadataStorage } from "../metadata/getMetadataStorage";
 import { getNameDecoratorParams } from "../helpers/decorators";
 import { DescriptionOptions, AbstractClassOptions, ResolveTypeOptions } from "./types";
 
-export type InterfaceOptions = DescriptionOptions & AbstractClassOptions & ResolveTypeOptions;
+export type InterfaceTypeOptions = DescriptionOptions & AbstractClassOptions & ResolveTypeOptions;
 
 export function InterfaceType(): ClassDecorator;
-export function InterfaceType(options: InterfaceOptions): ClassDecorator;
-export function InterfaceType(name: string, options?: InterfaceOptions): ClassDecorator;
+export function InterfaceType(options: InterfaceTypeOptions): ClassDecorator;
+export function InterfaceType(name: string, options?: InterfaceTypeOptions): ClassDecorator;
 export function InterfaceType(
-  nameOrOptions?: string | InterfaceOptions,
-  maybeOptions?: InterfaceOptions,
+  nameOrOptions?: string | InterfaceTypeOptions,
+  maybeOptions?: InterfaceTypeOptions,
 ): ClassDecorator {
   const { name, options } = getNameDecoratorParams(nameOrOptions, maybeOptions);
   return target => {